requirements.txt
bert-extractive-summarizer==0.10.1
torch==2.1.1
streamlit==1.29.0


